requirements.txt
streamlit==1.29.0
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
requests==2.31.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
loguru==0.7.2
tenacity==8.2.3
aiohttp==3.9.1
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0